Map `long` and `Long` arguments to `BIGINT`
The SQL type `INTEGER` often (always?) maps to integer types which are only 32 bits (4 bytes) long while `BIGINT` maps to integer types which are 64 bits (8 bytes) long.

Considering this, mapping `long` and `Long` to `BIGINT` instead of `INTEGER` seems to make more sense.

FWIW, this is motivated by a bug we discovered in CockroachDB when sending `NULL` values which were sent as `NULL::INT4` by Jdbi 3.20.1 instead of `NULL::INT8` for a column declared as `INT8`.
https://github.com/cockroachdb/cockroach/issues/67605

+++ core/src/main/java/org/jdbi/v3/core/argument/BoxedArgumentFactory.java
@@ -23,7 +23,7 @@ class BoxedArgumentFactory extends DelegatingArgumentFactory {
         register(Character.class, Types.CHAR, new ToStringBinder<>(PreparedStatement::setString));
         register(Short.class, Types.SMALLINT, PreparedStatement::setShort);
         register(Integer.class, Types.INTEGER, PreparedStatement::setInt);
-        register(Long.class, Types.INTEGER, PreparedStatement::setLong);
+        register(Long.class, Types.BIGINT, PreparedStatement::setLong);
         register(Float.class, Types.FLOAT, PreparedStatement::setFloat);
         register(Double.class, Types.DOUBLE, PreparedStatement::setDouble);
     }
